refactor: remove limits.h and replace its macros with stdint.h ones

+++ src/macros.h
@@ -10,7 +10,7 @@
 
 #define IN_RANGE(n,lo,hi) (((n) > (lo)) && ((n) < (hi)))
 
-#define DOUBLE_IS_INTEGER(d) ((d) <= LLONG_MIN || (d) >= LLONG_MAX || (d) == (long long)(d))
+#define DOUBLE_IS_INTEGER(d) ((d) <= INT64_MIN || (d) >= INT64_MAX || (d) == (int64_t)(d))
 
 #endif
